backend/requirements.txt
fastapi
uvicorn
openai
pymupdf4llm
python-dotenv
langchain-text-splitters
numpy
